Checkers Game With Obstacles
Checkers Game is an AI based board game using the rules of commonly
known Checkers game. The goal of players in game is to move the pieces
in the right corner of the screen to the other corner with the shortest
number of moves as possible. The role of the AI was to determine number
of shortest possible moves for each given board configuration. It can be
played by either user or computer. It’s developed as a mobile game for
Android platform using the Java language.